refactor tests to run in parallel, create framework and test packages, remove utils package
