+++ adapters/services/virus_total_test.go
@@ -3,18 +3,18 @@ package services
 import (
 	"github.com/joho/godotenv"
 	"github.com/simplycubed/vulnscan/entities"
-	"github.com/simplycubed/vulnscan/utils"
+	"github.com/simplycubed/vulnscan/test"
 	"github.com/stretchr/testify/assert"
 	"os"
 	"testing"
 )
 
 func TestVirusTotalAdapter(t *testing.T) {
-	mainFolder, _ := utils.FindMainFolder()
+	mainFolder, _ := test.FindMainFolder()
 	assert.NoError(t, godotenv.Load(mainFolder+string(os.PathSeparator)+".env"))
 	apiKey := os.Getenv("VIRUS_TOTAL_API_KEY")
 	assert.NotEmpty(t, apiKey)
-	path, _ := utils.FindTest("apps", "binary.ipa")
+	path, _ := test.FindTest("adapters", "services", "binary.ipa")
 	var (
 		command = entities.Command{VirusTotalKey: apiKey, Path: path}
 		entity  entities.VirusAnalysis
@@ -27,7 +27,7 @@ func TestVirusTotalAdapter(t *testing.T) {
 
 
 func TestHashMD5(t *testing.T) {
-	file, _ := utils.FindTest("apps", "binary.ipa")
+	file, _ := test.FindTest("adapters", "services", "binary.ipa")
 	if hash, e := hashMD5(file); e != nil {
 		t.Error(e)
 	} else if len(hash) != 32 {
