+++ adapters/tools/class_dump.go
@@ -5,12 +5,13 @@ import (
 	"github.com/simplycubed/vulnscan/entities"
 	"os"
 	"os/exec"
+	"path/filepath"
 )
 
 func ClassDumpAdapter(command entities.Command, entity *entities.BinaryAnalysis) error {
-	commandBin := command.Tools + "class-dump-z"
+	commandBin := filepath.Join(command.Tools, "class-dump-z")
 	if entity.BinType == entities.Swift {
-		commandBin = command.Tools + "class-dump-swift"
+		commandBin = filepath.Join(command.Tools, "class-dump-swift")
 	}
 	if _, err := os.Stat(commandBin); os.IsNotExist(err) {
 		return fmt.Errorf("class dump binary not found on %s, probably it's not installed", commandBin)
