+++ adapters/tools/common.go
@@ -5,6 +5,7 @@ import (
 	"github.com/simplycubed/vulnscan/entities"
 	"os"
 	"os/exec"
+	"path/filepath"
 	"regexp"
 	"strings"
 )
@@ -12,9 +13,9 @@ import (
 
 
 func performJtoolAnalysis(command entities.Command, args [][]string) (out string, err error) {
-	com := command.Tools + "jtool"
+	com := filepath.Join(command.Tools, "jtool")
 	if _, err := os.Stat(com); os.IsNotExist(err) {
-		return out, fmt.Errorf("jtool not found on %s, probably it's not installed", command.Path)
+		return out, fmt.Errorf("jtool not found on %s, probably it's not installed", command.Tools)
 	}
 	var sb strings.Builder
 	for _, arg := range args {
