+++ adapters/tools/downloader_test.go
@@ -1,28 +1,29 @@
 package tools
 
 import (
-	"fmt"
 	"github.com/simplycubed/vulnscan/entities"
-	"github.com/simplycubed/vulnscan/utils"
+	"github.com/simplycubed/vulnscan/test"
+	"github.com/simplycubed/vulnscan/framework"
 	"github.com/stretchr/testify/assert"
 	"os"
+	"path/filepath"
 	"testing"
 )
 
 
 func TestDownloaderAdapter(t *testing.T) {
 	var (
-		p, _ = utils.FindTest()
+		p, _ = test.FindTest()
 		command = entities.Command{Tools: p}
 		entity = entities.ToolUrls{
-			JTool:          utils.JtoolUrl,
-			ClassDumpZ:     utils.ClassDumpZUrl,
-			ClassDumpSwift: utils.ClassDumpSwiftUrl,
+			JTool:          framework.JtoolUrl,
+			ClassDumpZ:     framework.ClassDumpZUrl,
+			ClassDumpSwift: framework.ClassDumpSwiftUrl,
 		}
 	)
 	assert.NoError(t, DownloaderAdapter(command, &entity))
 	for k, _ := range entity.ToMap() {
-		tPath := fmt.Sprintf("%s/%s", p, k)
+		tPath := filepath.Join(p, k)
 		_, err := os.Stat(tPath)
 		assert.False(t, os.IsNotExist(err))
 		err = os.Remove(tPath)
