+++ entities/tool_urls.go
@@ -14,14 +14,14 @@ type ToolUrls struct {
 
 func (e *ToolUrls) ToMap() map[string]interface{} {
 	return map[string]interface{}{
-		"j_tool": e.JTool,
-		"class_dump_z": e.ClassDumpZ,
-		"class_dump_swift": e.ClassDumpSwift,
+		"jtool": e.JTool,
+		"class-dump-z": e.ClassDumpZ,
+		"class-dump-swift": e.ClassDumpSwift,
 	}
 }
 
 func (e *ToolUrls) FromMap(m map[string]interface{}) (ent Entity, err error) {
-	if v, ok := m["class_dump_z"]; ok {
+	if v, ok := m["class-dump-z"]; ok {
 		switch v.(type) {
 		case string:
 			e.ClassDumpZ = v.(string)
@@ -29,7 +29,7 @@ func (e *ToolUrls) FromMap(m map[string]interface{}) (ent Entity, err error) {
 			return ent, fmt.Errorf("erroneus class_dump_z url type, expected string, found: %T", v)
 		}
 	}
-	if v, ok := m["class_dump_swift"]; ok {
+	if v, ok := m["class-dump-swift"]; ok {
 		switch v.(type) {
 		case string:
 			e.ClassDumpSwift = v.(string)
@@ -37,7 +37,7 @@ func (e *ToolUrls) FromMap(m map[string]interface{}) (ent Entity, err error) {
 			return ent, fmt.Errorf("erroneus class_dump_swift url type, expected string, found: %T", v)
 		}
 	}
-	if v, ok := m["j_tool"]; ok {
+	if v, ok := m["jtool"]; ok {
 		switch v.(type) {
 		case string:
 			e.JTool = v.(string)
