+++ framework/tools_urls.go
@@ -1,0 +1,8 @@
+package framework
+
+// Urls to download the tolls
+const (
+	JtoolUrl = "https://github.com/simplycubed/vulnscan-dependencies/releases/download/0.0.1-beta/jtool"
+	ClassDumpZUrl = "https://github.com/simplycubed/vulnscan-dependencies/releases/download/0.0.1-beta/class-dump-z"
+	ClassDumpSwiftUrl = "https://github.com/simplycubed/vulnscan-dependencies/releases/download/0.0.1-beta/class-dump-swift"
+)
