+++ test/data/configuration/vulnscan.toml
@@ -1,0 +1,7 @@
+ 
+scans = ["binary", "code", "plist"]
+json = true
+source = "/Users/josemanuelgonzalezdelcampo/Proyectos/UpWork/vulnscan/test_files/apps/source.zip"
+binary = "/Users/josemanuelgonzalezdelcampo/Proyectos/UpWork/vulnscan/test_files/apps/binary.ipa"
+virus = "virus_scan_password"
+country = "us"
