--- test_files/configuration/vulnscan.toml
@@ -1,7 +1,0 @@
- 
-scans = ["binary", "code", "plist"]
-json = true
-source = "/Users/josemanuelgonzalezdelcampo/Proyectos/UpWork/vulnscan/test_files/apps/source.zip"
-binary = "/Users/josemanuelgonzalezdelcampo/Proyectos/UpWork/vulnscan/test_files/apps/binary.ipa"
-virus = "virus_scan_password"
-country = "us"
