--- test_files/plist/plist1.plist
@@ -1,45 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-	<key>CFBundleExecutable</key> <string>MoreSettings</string>
-
-	<key>CFBundleIconFiles</key>
-	<array>
-		<string>Icon.png</string>
-		<string>Icon@2x.png</string>
-		<string>Icon@3x.png</string>
-	</array>
-	<key>CFBundleIcons</key>
-	<dict>
-		<key>CFBundlePrimaryIcon</key>
-		<dict>
-			<key>CFBundleIconFiles</key>
-			<array>
-				<string>Icon.png</string>
-				<string>Icon@2x.png</string>
-				<string>Icon@3x.png</string>
-			</array>
-			<key>UIPrerenderedIcon</key> <string>true</string>
-		</dict>
-	</dict>
-
-	<key>CFBundleIdentifier</key> <string>org.h6nry.moresettings</string>
-	<key>CFBundleInfoDictionaryVersion</key> <string>6.0</string>
-	<key>CFBundlePackageType</key> <string>APPL</string>
-	<key>CFBundleSignature</key> <string>????</string>
-	<key>CFBundleSupportedPlatforms</key>
-	<array>
-		<string>iPhoneOS</string>
-	</array>
-	<key>CFBundleVersion</key> <string>1.0</string>
-	<key>DTPlatformName</key> <string>iphoneos</string>
-	<key>DTSDKName</key> <string>iphoneos3.0</string>
-	<key>LSRequiresIPhoneOS</key> <string>1</string>
-	<key>MinimumOSVersion</key> <string>3.0</string>
-	<key>UIBackgroundModes</key>
-	<array>
-		<string>audio</string>
-	</array>
-</dict>
-</plist>
