+++ usecases/store/store_analysis_test.go
@@ -4,7 +4,7 @@ import (
 	"encoding/json"
 	"github.com/simplycubed/vulnscan/adapters/mocks"
 	"github.com/simplycubed/vulnscan/entities"
-	"github.com/simplycubed/vulnscan/utils"
+	"github.com/simplycubed/vulnscan/test"
 	"github.com/stretchr/testify/assert"
 	"io/ioutil"
 	"testing"
@@ -12,7 +12,7 @@ import (
 
 func storeTestAdapter(command entities.Command, entity entities.Entity) error {
 	ent := entity.(*entities.StoreAnalysis)
-	sampleStoreResultPath, _ := utils.FindTest("sample-store-lookup.json")
+	sampleStoreResultPath, _ := test.FindTest("usecases", "store", "sample-store-lookup.json")
 	sampleLookupFile, _ := ioutil.ReadFile(sampleStoreResultPath)
 	comparisonEntity := entities.StoreAnalysis{}
 	assert.NoError(command.T, json.Unmarshal(sampleLookupFile, &comparisonEntity))
