--- utils/path.go
@@ -1,101 +1,0 @@
-package utils
-
-import (
-	"bytes"
-	"fmt"
-	"io"
-	"log"
-	"os"
-	"path/filepath"
-	"strings"
-)
-
-
-// FindMainFolder finds the main project folder
-func FindMainFolder() (string, error) {
-	dir, err := os.Getwd()
-	if err != nil {
-		log.Fatal(err)
-	}
-	absPath, _ := filepath.Abs(dir)
-	pathList := strings.Split(absPath, string(os.PathSeparator))
-	// We search for the last "vulnscan" in the current path, that should be the project main folder.
-	if pos := func(s []string) int {
-		last := -1
-		for p, v := range s {
-			if v == "vulnscan" {
-				last = p
-			}
-		}
-		return last
-	}(pathList); pos != len(pathList)-1 {
-		if pos == -1 {
-			// If we don't find the
-			return "", fmt.Errorf("vulnscan main directory not found")
-		}
-		// If we are not in the vulnscan folder, we remove all the children from the pathList
-		pathList = pathList[:pos+1]
-	}
-	return strings.Join(pathList, string(os.PathSeparator)), nil
-}
-
-// FindTest is a utility function to help find the test_files folder in different test environments. This solves
-// the problem where running tests from an IDE and from command line differed in how the test_files
-// folder was determined to be.
-//
-// The function first looks for an environment variable with the project path. If it doesn't exist, it tries
-// to guess it looking for the vulnscan folder in the current absolute path.
-func FindTest(testFilePath ...string) (string, error) {
-	var sb strings.Builder
-	if envPath := os.Getenv("TEST_WORKING_DIRECTORY"); len(envPath) > 0 {
-		sb.WriteString(envPath)
-		if !strings.HasPrefix(envPath, string(os.PathSeparator)) {
-			sb.WriteString(string(os.PathSeparator))
-		}
-		sb.WriteString("test_files")
-		for _, s := range testFilePath {
-			sb.WriteString(string(os.PathSeparator))
-			sb.WriteString(s)
-		}
-		return sb.String(), nil
-	}
-	mainFolder, err := FindMainFolder()
-	if err != nil {
-		return "", err
-	}
-	sb.WriteString(mainFolder)
-	sb.WriteString(string(os.PathSeparator))
-	sb.WriteString("test_files")
-	for _, s := range testFilePath {
-		sb.WriteString(string(os.PathSeparator))
-		sb.WriteString(s)
-	}
-	return sb.String(), nil
-}
-
-// WithPipeStdout writes to standard out
-func WithPipeStdout(printerFunc func() error) (string, error) {
-	// Let's hack the stdout to get the help message captured
-	old := os.Stdout
-	r, w, _ := os.Pipe()
-	os.Stdout = w
-
-	// Use a goroutine so printing can't block indefinitely
-	outC := make(chan string)
-	go func() {
-		var buf bytes.Buffer
-		_, _ = io.Copy(&buf, r)
-		outC <- buf.String()
-	}()
-
-	// Generate the help message
-	if e := printerFunc(); e != nil {
-		return "", e
-	}
-
-	// Collect the results (the regex is needed to remove some noise generated by the context)
-	_ = w.Close()
-	os.Stdout = old
-
-	return <-outC, nil
-}
